api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0
requests==2.32.3
ffmpeg-python==0.2.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
